+++ Gw2Sharp/WebApi/Http/HttpResponseInfo.cs
@@ -4,7 +4,6 @@ using System.Globalization;
 using System.Net;
 using System.Net.Http.Headers;
 using Gw2Sharp.Extensions;
-using Gw2Sharp.WebApi.Middleware;
 
 namespace Gw2Sharp.WebApi.Http
 {
@@ -14,16 +13,12 @@ namespace Gw2Sharp.WebApi.Http
     public class HttpResponseInfo
     {
         /// <summary>
-        /// The custom Gw2Sharp cache state header that may be injected by <see cref="CacheMiddleware"/>.
-        /// </summary>
-        public const string CACHE_STATE_HEADER = "X-Gw2Sharp-Cache-State";
-
-        /// <summary>
         /// Initializes a new instance of the <see cref="HttpResponseInfo"/> class with a <see cref="IWebApiResponse{T}"/> as base.
         /// </summary>
         /// <param name="statusCode">The HTTP status code.</param>
+        /// <param name="cacheState">The cache state.</param>
         /// <param name="responseHeaders">The HTTP response headers.</param>
-        public HttpResponseInfo(HttpStatusCode statusCode, IReadOnlyDictionary<string, string>? responseHeaders)
+        public HttpResponseInfo(HttpStatusCode statusCode, CacheState cacheState, IReadOnlyDictionary<string, string>? responseHeaders)
         {
             responseHeaders ??= new Dictionary<string, string>();
 
@@ -34,7 +29,7 @@ namespace Gw2Sharp.WebApi.Http
             this.LastModified = ParseResponseHeader(responseHeaders, "Last-Modified", ParseNullableDateTime);
             this.CacheMaxAge = ParseResponseHeader(responseHeaders, "Cache-Control", ParseNullableMaxAgeCache);
             this.Expires = ParseResponseHeader(responseHeaders, "Expires", ParseNullableDateTime);
-            this.CacheState = ParseResponseHeader(responseHeaders, CACHE_STATE_HEADER, ParseEnum<CacheState>);
+            this.CacheState = cacheState;
 
             static DateTimeOffset? ParseNullableDateTime(string value) =>
                 !string.IsNullOrWhiteSpace(value) ? (DateTimeOffset?)DateTimeOffset.Parse(value, CultureInfo.InvariantCulture) : null;
@@ -47,9 +42,6 @@ namespace Gw2Sharp.WebApi.Http
                     value = value[1..^1];
                 return CacheControlHeaderValue.Parse(value).MaxAge;
             }
-
-            static T ParseEnum<T>(string value) where T : struct =>
-                Enum.TryParse<T>(value, out var @enum) ? @enum : default;
         }